Factor components of functions module
The functions module contains impl blocks for both the
Session and Pcks11 structs. This commit moves those blocks
into the respective modules where those structs are defined.
Files session_management.rs and slot_token_mangement.rs
contained impl blocks for both, and were split in the process.
