+++ cryptoki/src/context/session_management.rs
@@ -1,0 +1,30 @@
+// Copyright 2021 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
+//! Session management functions
+
+use crate::context::Pkcs11;
+use crate::error::{Result, Rv};
+use crate::session::{Session, SessionFlags};
+use crate::slot::Slot;
+use std::convert::TryInto;
+
+impl Pkcs11 {
+    /// Open a new session with no callback set
+    pub fn open_session_no_callback(&self, slot_id: Slot, flags: SessionFlags) -> Result<Session> {
+        let mut session_handle = 0;
+
+        unsafe {
+            Rv::from(get_pkcs11!(self, C_OpenSession)(
+                slot_id.try_into()?,
+                flags.into(),
+                // TODO: abstract those types or create new functions for callbacks
+                std::ptr::null_mut(),
+                None,
+                &mut session_handle,
+            ))
+            .into_result()?;
+        }
+
+        Ok(Session::new(session_handle, self))
+    }
+}
