--- cryptoki/src/functions/mod.rs
@@ -1,13 +1,0 @@
-// Copyright 2021 Contributors to the Parsec project.
-// SPDX-License-Identifier: Apache-2.0
-//! PKCS11 functions
-
-pub(crate) mod decryption;
-pub(crate) mod encryption;
-pub(crate) mod general_purpose;
-pub(crate) mod key_management;
-pub(crate) mod object_management;
-pub(crate) mod random;
-pub(crate) mod session_management;
-pub(crate) mod signing_macing;
-pub(crate) mod slot_token_management;
